% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Multivariate Normal Sampler}
\usage{
mvrnormArma(mu, sigma)
}
\arguments{
\item{mu}{Mean of the multivariate normal distribution (has to be specified as a matrix with
1 row and number of columns equal to the number of variables in the multivariate distribution e.g. 2 for bivariate)}

\item{sigma}{A covariance matrix associated with mu}
}
\description{
Random number generation for a multivariate normal (Gaussian) distribution. Uses a Cholesky
decomposition to propose new values from a multivariate normal distribution with a given mean
vector (mu) and a specified covariance matrix (sigma).
}
\details{
Note: Currently (for my own modelling purposes) won't return values < 0.
}
